backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pdfplumber
python-docx
python-pptx
openpyxl
requests
beautifulsoup4
readability-lxml
google-cloud-storage
google-auth
python-dotenv
psycopg2-binary
bcrypt
pyjwt
python-jose[cryptography]
nltk
tiktoken
faiss-cpu
reportlab
pandas
tabulate
google-cloud-secret-manager
google-cloud-logging
google-cloud-monitoring
google-api-python-client
Pillow
openai
mistralai
httpx
python-multipart
redis
neo4j
pyotp
qrcode[pil]